Revert "2.15.42"
This reverts commit 310531efdbef77616935c798ea410ac196dfb203.

+++ binancefutures-api/src/main/java/io/contek/invoker/binancefutures/api/websocket/user/AccountUpdateChannel.java
@@ -5,6 +5,7 @@ import io.contek.invoker.binancefutures.api.websocket.common.WebSocketEventMessa
 import javax.annotation.concurrent.Immutable;
 import javax.annotation.concurrent.NotThreadSafe;
 import javax.annotation.concurrent.ThreadSafe;
+import java.math.BigDecimal;
 import java.util.List;
 
 import static io.contek.invoker.binancefutures.api.websocket.user.constants.UserEventTypeKeys._ACCOUNT_UPDATE;
@@ -51,20 +52,20 @@ public final class AccountUpdateChannel
   public static final class BalanceUpdate {
 
     public String a; // asset
-    public Double wb; // wallet balance
-    public Double cw; // cross wallet balance
+    public BigDecimal wb; // wallet balance
+    public BigDecimal cw; // cross wallet balance
   }
 
   @NotThreadSafe
   public static final class PositionUpdate {
 
     public String s; // symbol
-    public Double pa; // position amount
-    public Double ep; // entry price
-    public Double cr; // pre-fee accumulated realized
-    public Double up; // unrealized PnL
+    public BigDecimal pa; // position amount
+    public BigDecimal ep; // entry price
+    public BigDecimal cr; // pre-fee accumulated realized
+    public BigDecimal up; // unrealized PnL
     public String mt; // margin type
-    public Double iw; // isolated wallet
+    public BigDecimal iw; // isolated wallet
     public String ps; // position side
   }
 }
